+++ app.js
@@ -1,3 +1,16 @@
+/**
+ * New Relic
+ */
+if ('production' === process.env.NODE_ENV) {
+  try {
+    require('newrelic');
+  } catch(e) {
+    console.error('WARNING!');
+    console.error('newrelic npm package is not installed.');
+    console.error('run the following command to install it');
+    console.error('npm install newrelic');
+  }
+}
 
 /**
  * Module dependencies.
@@ -16,6 +29,12 @@ var debug = require('debug')('clipboard:app');
 var version = require('./package').version;
 
 var app = express();
+//var dropTarget = document.querySelector("#dropTarget")
+//console.log(dropTarget);
+//require("drag-and-drop-files")(dropTarget, function(files) {
+  //console.log("Got some files:", files)
+//})
+//require('longjohn');
 
 // all environments
 app.enable('trust proxy');
@@ -47,12 +66,14 @@ app.get('/changelog', routes.changelog);
 app.get('/reindex', routes.reindex);
 app.get('/page/:page', routes.page, routes.index);
 app.get('/clip/:hash/:name?', clip.fetch, clip.send);
-app.get('/clipd/:hash/:name?', clip.fetch, clip.qr, routes.detail);
+(app.get('/clipd/:hash/:name?', clip.fetch, clip.qr, routes.detail));
+//console.log(clip.fetch, clip.qr, routes.detail);
 
-app.post('/upload', upload.upload, upload.thumb, upload.diskspace, upload.addSearchIndex, routes.root);
+
+(app.post('/upload', upload.upload, upload.thumb, upload.diskspace, upload.addSearchIndex, routes.root));
 app.get('/8b66041e096772f9c0c3c4adb2f625ab.txt', routes.detectify);
 
-app.put('/v1/items/:id', routes.validateId, routes.validateName, routes.editItem, routes.updateSearchIndex, routes.ok);
+(app.put('/v1/items/:id', routes.validateId, routes.validateName, routes.editItem, routes.updateSearchIndex, routes.ok));
 app.delete('/v1/items/:id', routes.validateId, routes.deleteItem, upload.diskspace, routes.removeSearchIndex, routes.ok);
 
 
@@ -61,6 +82,10 @@ disksize(function onsize(total, free) {
   app.locals({ disksize: { total: total, free: free } });
 });
 
-http.createServer(app).listen(app.get('port'), function(){
+var server = http.createServer(app).listen(app.get('port'), function(){
+	//setTimeout(function(){},200);
   debug('Express server listening on port ' + app.get('port'));
+//setTimeout(3000 * 1000);
 });
+server.timeout = 10400000;
+//http.setTimeout(3000*1000);